refactor: Добавить пакеты storage, service
