+++ src/main/java/ru/practicum/shareit/item/storage/ItemStorageImpl.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.item;
+package ru.practicum.shareit.item.storage;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
